[FE] UPDATE : map 컴포넌트  CSS 적용
Former-commit-id: f8740f4bcd6f81b38a0bff7f41d0db718ad4d158

+++ frontend/bpservice/src/routes/auth/MyPage.js
@@ -9,10 +9,152 @@ import LoadingPage from "../../components/LoadingPage";
 import Nav from "../../components/Nav";
 import { deleteUser, getUserInfo } from "../../modules/mypage";
 
+import List from "@mui/material/List";
+import ListItem from "@mui/material/ListItem";
+import ListItemButton from "@mui/material/ListItemButton";
+import ListItemText from "@mui/material/ListItemText";
+import Box from "@mui/material/Box";
+import Divider from "@mui/material/Divider";
+import PersonIcon from "@mui/icons-material/Person";
+
 const myPageArea = css`
   width: 100%;
-  height: 72vh;
-  border: 1px black solid;
+  height: 69vh;
+`;
+
+const title = css`
+  flex-direction: row;
+  justify-content: space-around;
+  margin-left: 5vw;
+  margin-bottom: 1vh;
+`;
+
+const firstModalStyle = css`
+  width: 95vw;
+  margin-left: 2.5vw;
+  margin-bottom: 2.5vh;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  border-radius: 10px;
+  align-items: center;
+  background-color: rgba(249, 250, 251, 0.9);
+  box-shadow: 0 2px 6px rgba(0, 0, 0, 0.12), 0 2px 4px rgba(0, 0, 0, 0.24);
+  transition: all 0.3s cubic-bezier(0.25, 0.8, 0.25, 1);
+  .card-1:hover {
+    box-shadow: 0 14px 28px rgba(0, 0, 0, 0.55), 0 10px 10px rgba(0, 0, 0, 0.52);
+  }
+  input::placeholder {
+    color: transparent;
+  }
+  input:placeholder-shown + label {
+    color: #aaa;
+    font-size: 14pt;
+    top: 15px;
+  }
+  input:focus + label,
+  label {
+    color: #8aa1a1;
+    font-size: 10pt;
+    pointer-events: none;
+    position: absolute;
+    left: 0px;
+    top: 0px;
+    transition: all 0.2s ease;
+    -webkit-transition: all 0.2s ease;
+    -moz-transition: all 0.2s ease;
+    -o-transition: all 0.2s ease;
+  }
+  input:focus,
+  input:not(:placeholder-shown) {
+    border-bottom: solid 1px #8aa1a1;
+    outline: none;
+  }
+`;
+
+const secondModalStyle = css`
+  width: 95vw;
+  margin-left: 2.5vw;
+  display: flex;
+  margin-bottom: 2.5vh;
+  flex-direction: column;
+  justify-content: center;
+  border-radius: 10px;
+  align-items: center;
+  background-color: rgba(249, 250, 251, 0.9);
+  box-shadow: 0 2px 6px rgba(0, 0, 0, 0.12), 0 2px 4px rgba(0, 0, 0, 0.24);
+  transition: all 0.3s cubic-bezier(0.25, 0.8, 0.25, 1);
+  .card-1:hover {
+    box-shadow: 0 14px 28px rgba(0, 0, 0, 0.55), 0 10px 10px rgba(0, 0, 0, 0.52);
+  }
+  input::placeholder {
+    color: transparent;
+  }
+  input:placeholder-shown + label {
+    color: #aaa;
+    font-size: 14pt;
+    top: 15px;
+  }
+  input:focus + label,
+  label {
+    color: #8aa1a1;
+    font-size: 10pt;
+    pointer-events: none;
+    position: absolute;
+    left: 0px;
+    top: 0px;
+    transition: all 0.2s ease;
+    -webkit-transition: all 0.2s ease;
+    -moz-transition: all 0.2s ease;
+    -o-transition: all 0.2s ease;
+  }
+  input:focus,
+  input:not(:placeholder-shown) {
+    border-bottom: solid 1px #8aa1a1;
+    outline: none;
+  }
+`;
+
+const thirdModalStyle = css`
+  width: 95vw;
+  margin-left: 2.5vw;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  border-radius: 10px;
+  align-items: center;
+  background-color: rgba(249, 250, 251, 0.9);
+  box-shadow: 0 2px 6px rgba(0, 0, 0, 0.12), 0 2px 4px rgba(0, 0, 0, 0.24);
+  transition: all 0.3s cubic-bezier(0.25, 0.8, 0.25, 1);
+  .card-1:hover {
+    box-shadow: 0 14px 28px rgba(0, 0, 0, 0.55), 0 10px 10px rgba(0, 0, 0, 0.52);
+  }
+  input::placeholder {
+    color: transparent;
+  }
+  input:placeholder-shown + label {
+    color: #aaa;
+    font-size: 14pt;
+    top: 15px;
+  }
+  input:focus + label,
+  label {
+    color: #8aa1a1;
+    font-size: 10pt;
+    pointer-events: none;
+    position: absolute;
+    left: 0px;
+    top: 0px;
+    transition: all 0.2s ease;
+    -webkit-transition: all 0.2s ease;
+    -moz-transition: all 0.2s ease;
+    -o-transition: all 0.2s ease;
+  }
+  input:focus,
+  input:not(:placeholder-shown) {
+    border-bottom: solid 1px #8aa1a1;
+    outline: none;
+  }
 `;
 
 const sectionModalPosition = css`
@@ -24,16 +166,8 @@ const sectionModalPosition = css`
 `;
 
 const userInfoModal = css`
-  width: 80vw;
-  height: 10vh;
-  border: 1px black solid;
-  margin-bottom: 4vh;
   display: flex;
   justify-content: space-around;
-  .userInfoItem {
-    display: flex;
-    align-items: center;
-  }
 `;
 
 const contentModal = css`
@@ -63,15 +197,17 @@ function MyPage() {
   };
 
   const goToDeleteUser = () => {
-    dispatch(deleteUser());
+    if (!window.confirm("탈퇴를 하시겠습니까?")) {
+      return;
+    } else {
+      dispatch(deleteUser());
+    }
   };
 
   // 회원정보 - 아직 구현 X
   useEffect(() => {
     dispatch(getUserInfo());
   }, [dispatch]);
-
-  console.log(userInfo);
 
   return (
     <div>
@@ -81,39 +217,70 @@ function MyPage() {
 
       <Suspense fallback={<LoadingPage />}>
         <div css={myPageArea}>
-          <div css={sectionModalPosition}>
-            <div>
-              <h1>efs</h1>
-            </div>
-            {/* 유저 정보 */}
+          <h1 css={title}>My B.P</h1>
+
+          {/* 유저 정보 */}
+          <div css={firstModalStyle}>
             <div css={userInfoModal}>
-              <h1>{userName}</h1>
+              <div>
+                <PersonIcon></PersonIcon>
+              </div>
+              <div>
+                <h3>{userName}</h3>
+              </div>
               <div className="userInfoItem">
                 <span>{sido} </span>
                 <span>{sigungu} </span>
                 <span>{dong}</span>
               </div>
             </div>
-            <div css={contentModal}>
-              <div css={content}>
-                <h1>마이 페이지</h1>
+          </div>
 
-                {/* 회원 정보 수정 */}
-                <div>
-                  <button onClick={goToModifyInfo}>회원 정보 수정</button>
-                </div>
-
-                {/* 비밀번호 변경 */}
-                <div>
-                  <button onClick={goToModifyPwd}>비밀번호 변경</button>
-                </div>
-
-                {/* 회원 탈퇴 */}
-                <div>
-                  <button onClick={goToDeleteUser}>회원 탈퇴</button>
-                </div>
-              </div>
+          {/* 유저 정보 */}
+          <div css={secondModalStyle}>
+            <div css={content}>
+              <h1>마이 페이지</h1>
             </div>
+          </div>
+
+          {/* 회원 버튼 */}
+          <div css={thirdModalStyle}>
+            <Box
+              sx={{
+                width: "100%",
+                maxWidth: 360,
+                bgcolor: "background.paper",
+              }}
+            >
+              <nav aria-label="secondary mailbox folders">
+                <List>
+                  {/* 회원 정보 수정 */}
+                  <ListItem disablePadding>
+                    <ListItemButton onClick={goToModifyInfo}>
+                      <ListItemText primary="회원 정보 수정" />
+                    </ListItemButton>
+                  </ListItem>
+
+                  <Divider />
+
+                  <ListItem disablePadding>
+                    {/* 비밀번호 변경 */}
+                    <ListItemButton onClick={goToModifyPwd}>
+                      <ListItemText primary="비밀번호 변경" />
+                    </ListItemButton>
+                  </ListItem>
+
+                  <Divider />
+
+                  <ListItem disablePadding>
+                    {/* 회원 탈퇴 */}
+                    <ListItemButton onClick={goToDeleteUser}>
+                      <ListItemText primary="회원 탈퇴" />
+                    </ListItemButton>
+                  </ListItem>
+                </List>
+              </nav>
+            </Box>
           </div>
         </div>
       </Suspense>
